+++ R/misc_yhat.R
@@ -46,7 +46,8 @@ yhat.randomForest <- function(X.model, newdata, ...) {
     # if result is a vector then ncol parameter is null
     if (is.null(ncol(pred))) return(pred)
     #  binary classification
-    if (ncol(pred) == 2) return(pred[,2])
+    if (!is.null(attr(X.model, "predict_function_target_column"))) return(pred[,attr(X.model, "predict_function_target_column")])
+    if (ncol(pred) == 2 & is.null(attr(X.model, "predict_function_target_column"))) return(pred[,2])
 
   } else {
     pred <- predict(X.model, newdata, ...)
@@ -59,7 +60,9 @@ yhat.randomForest <- function(X.model, newdata, ...) {
 yhat.svm <- function(X.model, newdata, ...) {
   if (X.model$type == 0) {
     pred <- attr(predict(X.model, newdata = newdata, probability = TRUE), "probabilities")
-    if (ncol(pred) == 2) { # binary classification
+    if (!is.null(attr(X.model, "predict_function_target_column"))) { # binary classification
+      pred <- pred[,attr(X.model, "predict_function_target_column")]
+    } else if (ncol(pred) == 2 & is.null(attr(X.model, "predict_function_target_column"))) {
       pred <- pred[,2]
     }
   } else {
@@ -75,7 +78,11 @@ yhat.gbm <- function(X.model, newdata, ...) {
   response <- predict(X.model, newdata = newdata, n.trees = n.trees, type = "response")
   #gbm returns and 3D array for multilabel classif
   if(length(dim(response)) > 2){
-    response <- response[,,1]
+    if (!is.null(attr(X.model, "predict_function_target_column"))) {
+      response <- response[,,1][,attr(X.model, "predict_function_target_column")]
+    } else {
+      response <- response[,,1]
+    }
   }
   response
 }
@@ -102,7 +109,9 @@ yhat.cv.glmnet <- function(X.model, newdata, ...) {
     if (ncol(pred) == 1) {
       return(as.numeric(pred))
     }
-    if (ncol(pred) == 2) {
+    if (!is.null(attr(X.model, "predict_function_target_column"))) {
+      return(pred[,attr(X.model, "predict_function_target_column")])
+    } else if (ncol(pred) == 2 & is.null(attr(X.model, "predict_function_target_column"))) {
       return(pred[,2])
     }
   } else {
@@ -121,7 +130,11 @@ yhat.glmnet <- function(X.model, newdata, ...) {
     pred <- predict(X.model, newdata, type = "response", s = X.model$lambda[length(X.model$lambda)])
     #glmnet returns and 3D array for multilabel classif
     if(length(dim(pred)) > 2){
-      return(pred[,,1])
+      if (!is.null(attr(X.model, "predict_function_target_column"))) {
+        return(pred[,,1][,attr(X.model, "predict_function_target_column")])
+      } else {
+        return(pred[,,1])
+      }
     }
     # For binary classifiaction matrix with one column is returned
     if (ncol(pred) == 1) {
@@ -149,7 +162,11 @@ yhat.ranger <- function(X.model, newdata, ...) {
     # if result is a vector then ncol parameter is null
     if (is.null(ncol(pred))) return(pred)
     # binary classification
-    if (ncol(pred) == 2) return(pred[,2])
+    if (!is.null(attr(X.model, "predict_function_target_column"))) {
+      pred <- pred[,attr(X.model, "predict_function_target_column")]
+    } else if (ncol(pred) == 2 & is.null(attr(X.model, "predict_function_target_column"))) {
+      pred <- pred[, 2]
+    }
   }
   pred
 }
@@ -160,7 +177,9 @@ yhat.model_fit <- function(X.model, newdata, ...) {
   if (X.model$spec$mode == "classification") {
     response <- as.matrix(predict(X.model, newdata, type = "prob"))
     colnames(response) <- X.model$lvl
-    if (ncol(response) == 2) {
+    if (!is.null(attr(X.model, "predict_function_target_column"))) {
+      response <- response[,attr(X.model, "predict_function_target_column")]
+    } else if (ncol(response) == 2  & is.null(attr(X.model, "predict_function_target_column"))) {
       response <- response[,2]
     }
   }
@@ -176,7 +195,9 @@ yhat.model_fit <- function(X.model, newdata, ...) {
 yhat.train <- function(X.model, newdata, ...) {
   if (X.model$modelType == "Classification") {
     response <- predict(X.model, newdata = newdata, type = "prob")
-    if (ncol(response) == 2) {
+    if (!is.null(attr(X.model, "predict_function_target_column"))) {
+      response <- response[,attr(X.model, "predict_function_target_column")]
+    } else if (ncol(response) == 2 & is.null(attr(X.model, "predict_function_target_column"))) {
       response <- response[,2]
     }
   }
@@ -202,8 +223,10 @@ yhat.lrm <- function(X.model, newdata, ...) {
 yhat.rpart <- function(X.model, newdata, ...) {
   response <- predict(X.model, newdata = newdata)
   if (!is.null(dim(response))) {
-    if (ncol(response) == 2) {
-      response <- response[,2]
+    if (!is.null(attr(X.model, "predict_function_target_column"))) {
+      response <- response[,attr(X.model, "predict_function_target_column")]
+    } else if (ncol(response) == 2 & is.null(attr(X.model, "predict_function_target_column"))) {
+      response <- response[, 2]
     }
   }
   response
@@ -223,7 +246,9 @@ yhat.default <- function(X.model, newdata, ...) {
     return(as.numeric(response))
   }
   # result is a matrix of data.frame with a two column (binary classification), returns the second
-  if (ncol(response) == 2) {
+  if (!is.null(attr(X.model, "predict_function_target_column"))) {
+    return(as.numeric(response[,attr(X.model, "predict_function_target_column")]))
+  } else if (ncol(response) == 2 & is.null(attr(X.model, "predict_function_target_column"))) {
     return(as.numeric(response[,2]))
   }
   # result is a matrix of data.frame with more than 2 columns (multi label classification)